Реализовать проверку авторизации на бекэнде

+++ packages/server/sequelize/db-setup.ts
@@ -1,16 +1,20 @@
-import sequelize from './index';
-import { pickRandom } from '../utils/utils';
+import sequelize from './index'
+import { pickRandom } from '../utils/utils'
 
 export default async function setupDatabase() {
-  console.log('Will rewrite the SQLite example database, adding some dummy data.');
+  console.log(
+    'Will rewrite the SQLite example database, adding some dummy data.'
+  )
 
-  await sequelize.sync({ force: true });
+  await sequelize.models.message.sync({ force: true })
+  await sequelize.models.theme.sync({ force: true })
+  await sequelize.models.currentUser.sync({ force: true })
 
   await sequelize.models.theme.bulkCreate([
     { theme: 'How to become Junior' },
     { theme: 'How to become Middle' },
     { theme: 'How to become Senior' },
-  ]);
+  ])
 
   for (const theme of await sequelize.models.theme.findAll()) {
     for (let i = 0; i < 5; i++) {
@@ -21,7 +25,7 @@ export default async function setupDatabase() {
         'Mariya',
         'Alex',
         'Efim',
-      ]);
+      ])
 
       const message = pickRandom([
         'Hello!',
@@ -30,14 +34,17 @@ export default async function setupDatabase() {
         'Nice to meet you!',
         'Go to Stackoverflow',
         '42',
-        'Don\'t care',
-      ]);
+        "Don't care",
+      ])
 
-      // @ts-ignore
-      await sequelize.models.message.create({ message, author, themeId: theme.id });
+      await sequelize.models.message.create({
+        message,
+        author,
+        // @ts-ignore
+        themeId: theme.id,
+      })
     }
   }
 
-
-  console.log('Done!');
+  console.log('Done!')
 }
